+++ backend/src/main/java/com/dragonguard/backend/blockchain/exception/BlockchainException.java
@@ -1,4 +1,4 @@
-package com.dragonguard.backend.global.exception;
+package com.dragonguard.backend.blockchain.exception;
 
 public class BlockchainException extends IllegalStateException {
     private static final String MESSAGE = "블록체인 로직에 오류가 발생했습니다.";
