--- backend/src/main/java/com/dragonguard/backend/config/converter/ConverterConfig.java
@@ -1,14 +1,0 @@
-package com.dragonguard.backend.config.converter;
-
-import com.dragonguard.backend.search.converter.SearchTypeConverter;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.format.FormatterRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-public class ConverterConfig implements WebMvcConfigurer {
-    @Override
-    public void addFormatters(FormatterRegistry registry) {
-        registry.addConverter(new SearchTypeConverter());
-    }
-}
