--- backend/src/main/java/com/dragonguard/backend/search/converter/SearchTypeConverter.java
@@ -1,11 +1,0 @@
-package com.dragonguard.backend.search.converter;
-
-import com.dragonguard.backend.search.entity.SearchType;
-import org.springframework.core.convert.converter.Converter;
-
-public class SearchTypeConverter  implements Converter<String, SearchType> {
-    @Override
-    public SearchType convert(String source) {
-        return SearchType.valueOf(source.toUpperCase());
-    }
-}
